pwm: Enable frequencies in MHz range

+++ include/sc_pwm.h
@@ -31,7 +31,7 @@
 
 void sc_pwm_init(void);
 void sc_pwm_deinit(void);
-void sc_pwm_set_freq(uint16_t freq);
+void sc_pwm_set_freq(uint32_t freq);
 void sc_pwm_stop(int pwm);
 void sc_pwm_set_duty(int pwm, uint16_t duty);
